cleaned up folder structure
